refactor(model): reorganize animals into herbivore and predator subpackages
