+++ src/main/java/org/example/app/Main.java
@@ -4,7 +4,7 @@ import org.example.engine.IslandEngine;
 import org.example.map.IslandMap;
 
 public class Main {
-    private static final IslandMap map = new IslandMap(2, 2);
+    private static final IslandMap map = new IslandMap(10, 10);
 
     public static void main(String[] args) {
         new IslandInitializer(map).populate();
