+++ src/main/java/org/example/engine/IslandEngine.java
@@ -3,9 +3,9 @@ package org.example.engine;
 import org.example.map.Cell;
 import org.example.map.IslandMap;
 import org.example.model.animals.Animal;
-import org.example.model.animals.Herbivore;
+import org.example.model.animals.herbivore.Herbivore;
 import org.example.model.plants.Plant;
-import org.example.model.animals.Predator;
+import org.example.model.animals.predator.Predator;
 import org.example.statistics.StatisticPrinter;
 import org.example.statistics.StatisticTracker;
 import org.example.util.RandomUtil;
