--- src/main/java/org/example/model/animals/Bear.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Bear extends Predator{
-    public Bear() {
-        super(3, 80, 2, 5, 500);
-    }
-}
