--- src/main/java/org/example/model/animals/Boa.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Boa extends Predator{
-    public Boa() {
-        super(1, 3, 1, 30, 15);
-    }
-}
