--- src/main/java/org/example/model/animals/Boar.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Boar extends Herbivore{
-    public Boar() {
-        super(11, 50, 2, 50, 400);
-    }
-}
