--- src/main/java/org/example/model/animals/Buffalo.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Buffalo extends Herbivore{
-    public Buffalo() {
-        super(12, 100, 3, 10, 700);
-    }
-}
