--- src/main/java/org/example/model/animals/Caterpillar.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Caterpillar extends Herbivore{
-    public Caterpillar() {
-        super(14, 0, 0, 1000, 0.01);
-    }
-}
