--- src/main/java/org/example/model/animals/Deer.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Deer extends Herbivore{
-    public Deer() {
-        super(6, 50, 4, 20, 300);
-    }
-}
