--- src/main/java/org/example/model/animals/Duck.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Duck extends Herbivore{
-    public Duck() {
-        super(13, 0.15, 4, 200, 1);
-    }
-}
