--- src/main/java/org/example/model/animals/Eagle.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Eagle extends Predator{
-    public Eagle() {
-        super(4, 1, 3, 20, 6);
-    }
-}
