--- src/main/java/org/example/model/animals/Fox.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Fox extends Predator{
-    public Fox() {
-        super(2, 2, 2, 30, 8);
-    }
-}
