--- src/main/java/org/example/model/animals/Goat.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Goat extends Herbivore{
-    public Goat() {
-        super(9, 10, 3, 140, 60);
-    }
-}
