--- src/main/java/org/example/model/animals/Mouse.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Mouse extends Herbivore{
-    public Mouse() {
-        super(8, 0.01, 1, 500, 0.05);
-    }
-}
