--- src/main/java/org/example/model/animals/Sheep.java
@@ -1,7 +1,0 @@
-package org.example.model.animals;
-
-public class Sheep extends Herbivore{
-    public Sheep() {
-        super(10, 15, 3, 140, 70);
-    }
-}
