+++ src/main/java/org/example/model/animals/herbivore/Boar.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Boar extends Herbivore {
+    public Boar() {
+        super(11, 50, 2, 50, 400);
+    }
+}
