+++ src/main/java/org/example/model/animals/herbivore/Buffalo.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Buffalo extends Herbivore {
+    public Buffalo() {
+        super(12, 100, 3, 10, 700);
+    }
+}
