+++ src/main/java/org/example/model/animals/herbivore/Caterpillar.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Caterpillar extends Herbivore {
+    public Caterpillar() {
+        super(14, 0, 0, 1000, 0.01);
+    }
+}
