+++ src/main/java/org/example/model/animals/herbivore/Deer.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Deer extends Herbivore {
+    public Deer() {
+        super(6, 50, 4, 20, 300);
+    }
+}
