+++ src/main/java/org/example/model/animals/herbivore/Duck.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Duck extends Herbivore {
+    public Duck() {
+        super(13, 0.15, 4, 200, 1);
+    }
+}
