+++ src/main/java/org/example/model/animals/herbivore/Goat.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Goat extends Herbivore {
+    public Goat() {
+        super(9, 10, 3, 140, 60);
+    }
+}
