+++ src/main/java/org/example/model/animals/herbivore/Herbivore.java
@@ -1,4 +1,6 @@
-package org.example.model.animals;
+package org.example.model.animals.herbivore;
+
+import org.example.model.animals.Animal;
 
 public abstract class Herbivore extends Animal {
     public Herbivore(int id, double foodNeeded, int speed, int maxAmount, double weight) {
