+++ src/main/java/org/example/model/animals/herbivore/Mouse.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Mouse extends Herbivore {
+    public Mouse() {
+        super(8, 0.01, 1, 500, 0.05);
+    }
+}
