+++ src/main/java/org/example/model/animals/herbivore/Rabbit.java
@@ -1,4 +1,4 @@
-package org.example.model.animals;
+package org.example.model.animals.herbivore;
 
 public class Rabbit extends Herbivore {
     public Rabbit() {
