+++ src/main/java/org/example/model/animals/herbivore/Sheep.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.herbivore;
+
+public class Sheep extends Herbivore {
+    public Sheep() {
+        super(10, 15, 3, 140, 70);
+    }
+}
