+++ src/main/java/org/example/model/animals/predator/Bear.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.predator;
+
+public class Bear extends Predator {
+    public Bear() {
+        super(3, 80, 2, 5, 500);
+    }
+}
