+++ src/main/java/org/example/model/animals/predator/Boa.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.predator;
+
+public class Boa extends Predator {
+    public Boa() {
+        super(1, 3, 1, 30, 15);
+    }
+}
