+++ src/main/java/org/example/model/animals/predator/Eagle.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.predator;
+
+public class Eagle extends Predator {
+    public Eagle() {
+        super(4, 1, 3, 20, 6);
+    }
+}
