+++ src/main/java/org/example/model/animals/predator/Fox.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.predator;
+
+public class Fox extends Predator {
+    public Fox() {
+        super(2, 2, 2, 30, 8);
+    }
+}
