+++ src/main/java/org/example/model/animals/predator/Wolf.java
@@ -1,0 +1,7 @@
+package org.example.model.animals.predator;
+
+public class Wolf extends Predator {
+    public Wolf() {
+        super(0, 8, 3, 30, 50);
+    }
+}
